+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod amb;
 pub mod body;
 pub mod cli;
 pub mod client;